refactor(ui): add the UploadForm component
